Heyyyy, ban command made
Right, so last commit all over GitHub was 20th April so we are back. Most of the commands are prolly gonna be committed by me tho. I mean seeing this command there are like tons of things to improve but no we are too lazy to do it now. I believe we have time until 24th to complete it. Command made by @Renny01 btw!